copy and make blogLayout and modify layout
