+++ package.json
@@ -25,6 +25,7 @@
     "qs": "^6.9.4",
     "react": "16.13.1",
     "react-dom": "16.13.1",
+    "react-icons": "^3.10.0",
     "react-instantsearch-dom": "^6.6.0",
     "remark": "^12.0.0",
     "remark-breaks": "^1.0.5",
