+++ src/pages/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Head from 'next/head'
 import { Layout } from '../components/Layout'
-import { NewTop, About, History, Works } from '../components/HomeContents'
+import { Top, About, History, Works } from '../components/HomeContents'
 import blogConfig from '../../blog.config'
 
 export default function (props) {
@@ -17,7 +17,7 @@ export default function (props) {
           <meta property='og:image' content={`${blogConfig.baseUrl}/assets/prtsc700.jpg`} />
           <meta property='og:url' content={blogConfig.baseUrl} />
         </Head>
-        <NewTop/>
+        <Top/>
         <About />
         <History />
         <Works />
